Refactor LP dumping to dimod.lp namespace
Because CQM.from_lp_file() was deprecated as part of
https://github.com/dwavesystems/dimod/pull/1182
it makes sense to create LP files with lp.dump() and
lp.dumps() rather than CQM.to_lp_file().

This way we're not introducing a method only to
immediately deprecate it.
