+++ dimod/constrained.py
@@ -23,7 +23,6 @@ import copy
 import json
 import os.path
 import re
-import string
 import tempfile
 import uuid
 import warnings
@@ -1840,203 +1839,6 @@ class ConstrainedQuadraticModel:
 
         file.seek(0)
         return file
-
-    def to_lp_file(self, *, spool_size: int = int(1e9)) -> tempfile.SpooledTemporaryFile:
-        """Serialize to LP file format.
-
-        LP files are a common format for encoding optimization models. See
-        documentation from Gurobi_ and CPLEX_.
-
-        Args:
-            spool_size: Defines the `max_size` passed to the constructor of
-                :class:`tempfile.SpooledTemporaryFile`. Determines whether
-                the returned file-like's contents will be kept on disk or in
-                memory.
-
-        .. _Gurobi: https://www.gurobi.com/documentation/9.5/refman/lp_format.html
-
-        .. _CPLEX: https://www.ibm.com/docs/en/icos/12.8.0.0?topic=cplex-lp-file-format-algebraic-representation
-
-        """
-        LABEL_VALID_CHARS = set(string.ascii_letters + string.digits + "'!\"#$%&(),.;?@_‘’{}~")
-        LABEL_INVALID_FIRST_CHARS = set('eE.' + string.digits)
-
-        def sign(bias: float) -> str:
-            return '-' if bias < 0 else '+'
-
-        def abs_(bias: float) -> str:
-            # todo: use bias.is_integer() instead of int(bias) == bias once we can require
-            # numpy>=1.22.0.  this is needed to support Float32BQM on numpy<1.22.0.
-            return repr(abs(int(bias))) if int(bias) == bias else repr(abs(bias))
-
-        def sense(s: Sense) -> str:
-            return '=' if s.value == "==" else s.value
-
-        def label_error(label: Hashable, rule: str) -> str:
-            raise ValueError(f'Label {label!r} cannot be output to an LP file; {rule}. '
-                             'Use CQM.relabel_variables() or CQM.relabel_constraints() to fix')
-
-        def validate_label(label: Hashable):
-            if not isinstance(label, str):
-                label_error(label, 'labels must be strings')
-
-            if not len(label):
-                label_error(label, 'labels must be nonempty')
-
-            if len(label) > 255:
-                label_error(label, 'labels must be <= 255 characters in length')
-
-            if not all(c in LABEL_VALID_CHARS for c in label):
-                label_error(label, f'labels must be composed of only these characters: '
-                                   f'{"".join(sorted(LABEL_VALID_CHARS))}')
-
-            if any(label.startswith(c) for c in LABEL_INVALID_FIRST_CHARS):
-                label_error(label, f'labels must not start with any of these characters: '
-                                   f'{"".join(sorted(LABEL_INVALID_FIRST_CHARS))}')
-
-        class ToLPFile(SpooledTemporaryFile):
-            TARGET_LINE_LEN = 80
-
-            def __init__(self):
-                """Subclass :class:`SpooledTemporaryFile` to provide a custom `write` method that
-                inserts line breaks into the output stream.  This is done to generate attractive
-                output (e.g. lines that are not extremely short) for most inputs, while satisfying
-                the line length requirement of LP files for all inputs.
-
-                This design has the advantage of a simple and readable implementation as compared
-                to the approach of tracking line length and emitting newlines conditionally,
-                through additional statements interspersed with the f.write() calls below.  And it
-                has the advantage of producing more attractive output as compared to the dead
-                simple approach of emitting newlines unconditionally in the f.write() calls below.
-
-                This design does incur a run-time performance penalty of about 50% versus those
-                alternatives.  In the balance of our goals, this seems to be the best option.
-                """
-                super().__init__(max_size=spool_size, mode='wb')
-
-                # length of current line, i.e. number of bytes written since last newline
-                self._line_len = 0
-
-            def write(self, bytes_: ByteString) -> None:
-                """Check if writing ``bytes_`` would result in our line length target being
-                exceeded, and if so, emit a newline.  Then write ``bytes_``.
-
-                Note that this function does not gaurantee a maximum line length.  Lines longer
-                than the target length can be emitted.  Callers, i.e. the f.write() calls in the
-                code below, must pass ``bytes_`` arguments that are sufficiently short, to avoid
-                exceeding the maximum line length for LP files.  The longest lines emitted by the
-                current implementation are those encoding quadratic terms for an objective or
-                constraints, when two variables have maximum length labels.  We have a unit test
-                (test_long_line_breaking) to check this case.
-                """
-                try:
-                    pos = bytes_.index(b'\n')
-                    rpos = bytes_.rindex(b'\n')
-                except ValueError:
-                    pos = len(bytes_)
-                    rpos = None
-
-                if self._line_len + pos > self.TARGET_LINE_LEN - 1:
-                    super().write(b'\n ')
-                    self._line_len = 1
-
-                super().write(bytes_)
-
-                if rpos is not None:
-                    self._line_len = len(bytes_) - 1 - rpos
-                else:
-                    self._line_len += len(bytes_)
-
-        # check that constraint labels are serializable
-        for c in self.constraints:
-            validate_label(c)
-
-        vartypes = {var: self.vartype(var) for var in self.variables}
-
-        # check that variable labels and types are serializable
-        for v, vartype in vartypes.items():
-            validate_label(v)
-
-            if vartype == Vartype.SPIN:
-                raise ValueError(
-                    'SPIN variables not supported in LP files, convert them to BINARY beforehand.')
-
-        f = ToLPFile()
-
-        # write the objective
-        for var, bias in self.objective.linear.items():
-            if bias:
-                if not f.tell():
-                    # LP files allow to omit this part if the objective is empty, so
-                    # we add these lines if we are sure that there are nonzero terms.
-                    f.write(b"Minimize\n")
-                    f.write(b' obj: ')
-                f.write(f"{sign(bias)} {abs_(bias)} {var} ".encode('utf8'))
-
-        if self.objective.quadratic:
-            if not f.tell():
-                # if the objective is only quadratic then the header is written now
-                f.write(b"Minimize\n")
-                f.write(b' obj: ')
-
-            f.write(b'+ [ ')
-            for (u, v), bias in self.objective.quadratic.items():
-                # multiply bias by two because all quadratic terms are eventually
-                # divided by two outside the squared parenthesis
-                f.write(f"{sign(bias)} {abs_(2 * bias)} {u} * {v} ".encode('utf8'))
-            f.write(b']/2 ')
-
-        if self.objective.offset:
-            offset = self.objective.offset
-
-            if not f.tell():
-                # if the objective has only an offset then the header is written now
-                f.write(b"Minimize\n")
-                f.write(b' obj: ')
-            f.write(f"{sign(offset)} {abs_(offset)} ".encode('utf8'))
-
-        # write the constraints
-        f.write(b"\n\n")
-        f.write(b"Subject To \n")
-
-        for label, constraint in self.constraints.items():
-            f.write(f' {label}: '.encode('utf8'))
-
-            for var, bias in constraint.lhs.linear.items():
-                if bias:
-                    f.write(f"{sign(bias)} {abs_(bias)} {var} ".encode('utf8'))
-
-            if constraint.lhs.quadratic:
-                f.write(b'+ [ ')
-                for (u, v), bias in constraint.lhs.quadratic.items():
-                    f.write(f"{sign(bias)} {abs_(bias)} {u} * {v} ".encode('utf8'))
-                f.write(b'] ')
-
-            rhs = constraint.rhs - constraint.lhs.offset
-            f.write(f" {sense(constraint.sense)} {rhs}\n".encode('utf8'))
-
-        # write variable bounds
-        f.write(b'\n')
-        f.write(b'Bounds\n')
-
-        for v, vartype in vartypes.items():
-            if vartype is Vartype.INTEGER:
-                f.write(f' {self.lower_bound(v)} <= {v} <= {self.upper_bound(v)}\n'.encode('utf8'))
-
-        # write variable names
-        for section, vartype_ in (('Binary', Vartype.BINARY), ('General', Vartype.INTEGER)):
-            f.write(b'\n')
-            f.write(f'{section}\n'.encode('utf8'))
-
-            for v, vartype in vartypes.items():
-                if vartype is vartype_:
-                    f.write(f' {v}'.encode('utf8'))
-
-        # conclude
-        f.write(b'\n')
-        f.write(b'End')
-        f.seek(0)
-        return f
 
     def upper_bound(self, v: Variable) -> Bias:
         """Return the upper bound on the specified variable.
